Prepare for changes for configuration service (#186)
* Keep the extension view container object

Before this commit, the extension view container object was being
destroyed and created whenever the user moved from and then back to the
Extension Views.  This caused the equivalent of a refresh.  Now, the
object is created once and its display is changed whenever the user
moves from and to the Extension Views.

* Use classNames where appropriate

* Divert expected but undesirable error message

* Convert NavItem into a simple enum

* Add hasChatSupport to extension manifest types

* Improved isChatEnabled label

* Improve local-mode extension descriptions

* Show Project Overview when switching projects

* Correct action expected to display an extension

* Remove duplicate constants from run-list trigger

* Create Web socket on load instead of authorization

* Rename ExtensionFrameParams to ExtensionFrameOptions

This is in line with the naming in the Twitch coordinator.

* Correct asynchronous behavior of the Rig component

* Reduce URLs to paths in the API

* Replace the API object with a class

The enables more easily specifying whether the invocation is local or
online.  There are two instances provided for the majority of the cases
where it is known up-front which of the two invocation choices to use.

* Use a different endpont for fetchExtensionManifest

* Add zip function

This behaves much like the same-named function in Python.

* Update coordinator to 7.22.2

* Improved tests

* Use taskkill on Windows

This ensures all processes the front-end command creates are shut down.

+++ scripts/project.js
@@ -26,9 +26,7 @@ module.exports = function(app) {
       description: 'Front-end-only React-based example to get you started building quickly.',
       repository: 'twitchdev/extensions-boilerplate',
       frontendFolderName: 'extensions-boilerplate',
-      frontendCommand: process.platform === 'win32' ?
-        'node "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" run host' :
-        'node /usr/local/lib/node_modules/npm/bin/npm-cli.js run host',
+      frontendCommand: 'npm run host',
       backendCommand: '',
       npm: ['i'],
       sslFolderName: 'conf',
@@ -77,6 +75,9 @@ module.exports = function(app) {
           options.cwd = frontendFolderPath;
         } else if (projectFolderPath) {
           options.cwd = join(projectFolderPath, frontendFolderPath);
+        }
+        if (process.platform === 'win32') {
+          options.shell = true;
         }
       } else {
         args = [
@@ -192,7 +193,11 @@ module.exports = function(app) {
 
     async function stop(child) {
       if (child) {
-        child.kill();
+        if (process.platform === 'win32') {
+          childProcess.spawnSync('taskkill', ['/f', '/pid', child.pid, '/t']);
+        } else {
+          child.kill();
+        }
         return await new Promise((resolve, _) => {
           let hasResolved = false;
           child.stderr.on('data', (data) => process.stderr.write(data.toString()));
